CareBot/requirements.txt
Flask>=2.2.3
aiosqlite>=0.19.0
python-telegram-bot>=20.0
yoyo-migrations>=8.0.0
